==> search_engine_workshop/nb1_data_eda.ipynb <==
import polars as pl

import requests 

import pickle

docs_url = 'https://github.com/alexeygrigorev/llm-rag-workshop/raw/main/notebooks/documents.json'
docs_response = requests.get(docs_url)
documents_raw = docs_response.json()

type(documents_raw)

len(documents_raw)

type(documents_raw[0])

for key in documents_raw[0].keys():
    print(f'type({key}) = {type(documents_raw[0][key])}')

documents_raw[0]['course']

documents_raw[0]['documents'][0]

documents = []

for course in documents_raw:
    course_name = course['course']

    for doc in course['documents']:
        doc['course'] = course_name
        documents.append(doc)

len(documents)

documents[0]

df = pl.DataFrame(documents, schema=['course', 'section', 'question', 'text'])
df.head()

df['course'].value_counts(sort=True)

df_de = df.filter( pl.col('course').eq('data-engineering-zoomcamp') )
df_de.head()

df_de['section'].value_counts(sort=True)

df.write_parquet('df_qa.parquet')
with open("documents.pkl", "wb") as f:
    pickle.dump(documents, f)
