==> search_engine_workshop/nb2_vectorize.ipynb <==
import polars as pl
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity




df = pl.read_parquet('df_qa.parquet')
df = df.filter( pl.col('course').eq('data-engineering-zoomcamp') )
df.head()

## Vectorize Documents

- Turn emails into vectors
- term-document matrix
    - rows: documents
    - columns: words/tokens
- bag of words
    - word counts only

docs_example = [
    "January course details, register now",
    "Course prerequisites listed in January catalog",
    "Submit January course homework by end of month",
    "Register for January course, no prerequisites",
    "January course setup: Python and Google Cloud"
]

cv = CountVectorizer(stop_words='english')

cv.fit(docs_example)

X = cv.transform(docs_example)


print(cv.get_feature_names_out())

pl.from_numpy(np.array(X.todense()))

df_cv = pl.from_numpy(np.array(X.todense()) ,schema=cv.get_feature_names_out().tolist())
df_cv.transpose(include_header=True, header_name="Word")

### FROM BAG OF WORDS TO TFID


cv = TfidfVectorizer(stop_words='english')
X = cv.fit_transform(docs_example)

names = cv.get_feature_names_out()

df_docs = pl.DataFrame(X.toarray(), schema=list(names))

df_docs.with_columns(
    pl.selectors.float().round(2)
).transpose(include_header=True, header_name="Word")


query = "Do I need to know python to sign up for the January course?"



q = cv.transform([query])
q.toarray()



query_dict = dict(zip(names, q.toarray()[0]))
doc_dict = dict(zip(names, X.toarray()[1]))
print(query_dict)
print(doc_dict)

df_qd = pl.DataFrame([query_dict, doc_dict])
pl.concat([ pl.DataFrame({'index':['query','doc']}) ,df_qd],how='horizontal' )


print('query similarity to row 1')
df_qd.product().sum_horizontal().item()

print('query similarity to all docs')

X.dot(q.T).toarray()



print('equivalent to cosine similarity')

cosine_similarity(X, q)

print(query)

print(docs_example[4])
