==> search_engine_workshop/nb3_text_search.ipynb <==
import polars as pl
import numpy as np
import pickle
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

with open("documents.pkl", "rb") as f:
    documents = pickle.load(f)

df = pl.read_parquet('df_qa.parquet')
df = df.filter( pl.col('course').eq('data-engineering-zoomcamp') )
df.head()

df.columns

fields = ['section', 'question', 'text']
transformers = {}
matrices = {}

for field in fields:
    cv = TfidfVectorizer(stop_words='english', min_df=3)
    X = cv.fit_transform(df[field])

    transformers[field] = cv
    matrices[field] = X



transformers['text'].get_feature_names_out()

matrices['text']

435 documents with 1177 words tracked.

query = "I just signed up. Is it too late to join the course?"

Compare query to all known documents:

q = transformers['text'].transform([query])
score = cosine_similarity(matrices['text'], q).flatten()

print(score[:10])
print(np.mean(score))
print(np.median(score))
print(np.max(score))

idx = np.argsort(-score)[:10]
idx

score[idx]

print(query)

df.with_row_index().filter(pl.col("index").is_in(idx))['text'].to_list()


boost = {'question': 3.0}

score = np.zeros(len(df))

for f in fields:
    b = boost.get(f, 1.0)
    q = transformers[f].transform([query])
    s = cosine_similarity(matrices[f], q).flatten()
    score = score + b * s



idx = np.argsort(-score)[:10]
results = df.with_row_index().filter(pl.col("index").is_in(idx))
results.to_dicts()





class TextSearch:

    def __init__(self, text_fields):
        self.text_fields = text_fields
        self.matrices = {}
        self.vectorizers = {}

    def fit(self, records, vectorizer_params={}):
        self.df = pl.DataFrame(records)

        for f in self.text_fields:
            cv = TfidfVectorizer(**vectorizer_params)
            X = cv.fit_transform(self.df[f])
            self.matrices[f] = X
            self.vectorizers[f] = cv

    def search(self, query, n_results=10, boost={}, filters={}):
        score = np.zeros(len(self.df))

        for f in self.text_fields:
            b = boost.get(f, 1.0)
            q = self.vectorizers[f].transform([query])
            s = cosine_similarity(self.matrices[f], q).flatten()
            score = score + b * s

        for field, value in filters.items():
            s = self.df[field]
            mask = s.set(s != value,'0').set(s==value,'1').cast(pl.Int32).to_numpy()
            score = score * mask

        idx = np.argsort(-score)[:n_results]
        results = self.df.with_row_index().filter(pl.col("index").is_in(idx))
        return results.to_dicts()



index = TextSearch(text_fields=['section', 'question', 'text'])

index.fit(documents)

index.search(
    query='I just signed up. Is it too late to join the course?',
    n_results=5,
    boost={'question': 3.0},
    filters={'course': 'data-engineering-zoomcamp'}
)
